Replace inline script with Next.js Script component
